+++ rust/src/simple_state/find/search_state.rs
@@ -148,7 +148,7 @@ impl BarrieredKnowledge {
             status: FindStatus::Winning(move_),
         }
     }
-    pub fn new_losing<const MAX_CLIQUE: u32>(state: &State<MAX_CLIQUE>, depth: usize) -> Self {
+    pub fn new_losing<const MAX_CLIQUE: u32>(state: &State<MAX_CLIQUE>, depth: u16) -> Self {
         Self {
             barrier: state.limits_to_barrier(),
             status: FindStatus::Losing { depth },
@@ -196,7 +196,7 @@ impl BarrieredKnowledge {
 pub enum FindStatus {
     Winning(WinningMove),
     Losing {
-        depth: usize,
+        depth: u16,
     },
     InProgress,
     #[default]
@@ -204,12 +204,12 @@ pub enum FindStatus {
 }
 
 impl FindStatus {
-    fn success_key(&self) -> usize {
+    fn success_key(&self) -> u16 {
         match self {
             FindStatus::Winning(_) => 0,
-            FindStatus::Losing { depth } => usize::MAX - 1 - *depth,
-            FindStatus::InProgress => usize::MAX - 1,
-            FindStatus::Unknown => usize::MAX,
+            FindStatus::Losing { depth } => u16::MAX - 1 - *depth,
+            FindStatus::InProgress => u16::MAX - 1,
+            FindStatus::Unknown => u16::MAX,
         }
     }
     fn is_winning(&self) -> bool {
